Refactor events processor to make it extensible
